Добавлен функционал возвращения профилей

+++ controllers/api-applications-controller.js
@@ -2,7 +2,7 @@
 import {Application} from "../models/applications.js";
 
 const handleError = (res, error) => {
-    console.log('Something wrong: ', error);
+    console.log('Something went wrong!', error);
     res.status(500).send(error.message);
 };
 
@@ -48,12 +48,27 @@ export const postApplication = (req, res) => {
     // })
 
     Application
-        .create({dept, firstName, isComplete: false, isMinobr, lastName, middleName, org, phoneNumber, position, prevOrg, room})
+        .create({
+            dept,
+            firstName,
+            isComplete: false,
+            isMinobr,
+            lastName,
+            middleName,
+            org,
+            phoneNumber,
+            position,
+            prevOrg,
+            room
+        })
         .then((application) => res.status(200).json(application))
         .catch((error) => handleError(res, error))
 };
 
-// Контроллер получает часть апликейшена в запросе и обновляет апликейшен а БД, возвращает измененный апликейшен
+/** Контроллер получает часть апликейшена в запросе и обновляет апликейшен а БД, возвращает измененный апликейшен
+ * @param req Запрос с клиента
+ * @param res Ответ на клиент
+ */
 export const updateApplication = (req, res) => {
     const {isComplete} = req.body;
     const id = req.params.id
@@ -65,13 +80,14 @@ export const updateApplication = (req, res) => {
 
 /**
  * Метод находит в БД апликейшен по его ID и удаляет его, возвращает на клиент ID удаленного апликейшена
- * @param req - Запрос от клиента
- * @param res - Ответ на клиент
+ * @param req Запрос с клиента
+ * @param res Ответ на клиент
  */
 export const deleteApplication = (req, res) => {
-
-    Application
-        .findByIdAndDelete(req.params.id)
-        .then(() => res.status(200).json(req.params.id))
-        .catch((error) => handleError(res, error));
+    if (req.params.id.length > 1) {
+        Application
+            .findByIdAndDelete(req.params.id)
+            .then(() => res.status(200).json(req.params.id))
+            .catch((error) => handleError(res, error));
+    }
 };
